+++ lib/src/test/java/net/mm2d/upnp/SsdpNotifyReceiverTest.java
@@ -34,7 +34,7 @@ public class SsdpNotifyReceiverTest {
     public void setNotifyListener_受信メッセージが通知されること() throws Exception {
         final NetworkInterface networkInterface = NetworkUtils.getAvailableInet4Interfaces().get(0);
         final SsdpServerDelegate delegate = spy(new SsdpServerDelegate(mock(Receiver.class), Address.IP_V4, networkInterface));
-        final InterfaceAddress interfaceAddress = TestUtils.createInterfaceAddress("192.0.2.2", "255.255.255.0", (short) 16);
+        final InterfaceAddress interfaceAddress = TestUtils.createInterfaceAddress("192.0.2.2", "255.255.255.0", 16);
         doReturn(interfaceAddress).when(delegate).getInterfaceAddress();
         final SsdpNotifyReceiver receiver = spy(new SsdpNotifyReceiver(delegate));
 
@@ -53,7 +53,7 @@ public class SsdpNotifyReceiverTest {
     @Test
     public void onReceive_同一セグメントからのメッセージは通知する() throws Exception {
         final SsdpNotifyReceiver receiver = spy(new SsdpNotifyReceiver(Address.IP_V4, NetworkUtils.getAvailableInet4Interfaces().get(0)));
-        final InterfaceAddress address = TestUtils.createInterfaceAddress("192.0.2.1", "255.255.0.0", (short) 24);
+        final InterfaceAddress address = TestUtils.createInterfaceAddress("192.0.2.1", "255.255.0.0", 24);
         doReturn(address).when(receiver).getInterfaceAddress();
         final NotifyListener listener = mock(NotifyListener.class);
         receiver.setNotifyListener(listener);
@@ -67,7 +67,7 @@ public class SsdpNotifyReceiverTest {
     @Test
     public void onReceive_Listenerがnullでもクラッシュしない() throws Exception {
         final SsdpNotifyReceiver receiver = spy(new SsdpNotifyReceiver(Address.IP_V4, NetworkUtils.getAvailableInet4Interfaces().get(0)));
-        final InterfaceAddress address = TestUtils.createInterfaceAddress("192.0.2.1", "255.255.0.0", (short) 24);
+        final InterfaceAddress address = TestUtils.createInterfaceAddress("192.0.2.1", "255.255.0.0", 24);
         doReturn(address).when(receiver).getInterfaceAddress();
         final byte[] data = TestUtils.getResourceAsByteArray("ssdp-notify-alive0.bin");
 
@@ -77,7 +77,7 @@ public class SsdpNotifyReceiverTest {
     @Test
     public void onReceive_異なるセグメントからのメッセージは無視する() throws Exception {
         final SsdpNotifyReceiver receiver = spy(new SsdpNotifyReceiver(Address.IP_V4, NetworkUtils.getAvailableInet4Interfaces().get(0)));
-        final InterfaceAddress address = TestUtils.createInterfaceAddress("192.0.2.1", "255.255.0.0", (short) 24);
+        final InterfaceAddress address = TestUtils.createInterfaceAddress("192.0.2.1", "255.255.0.0", 24);
         doReturn(address).when(receiver).getInterfaceAddress();
         final NotifyListener listener = mock(NotifyListener.class);
         receiver.setNotifyListener(listener);
@@ -91,7 +91,7 @@ public class SsdpNotifyReceiverTest {
     @Test
     public void onReceive_M_SEARCHパケットは無視する() throws Exception {
         final SsdpNotifyReceiver receiver = spy(new SsdpNotifyReceiver(Address.IP_V4, NetworkUtils.getAvailableInet4Interfaces().get(0)));
-        final InterfaceAddress address = TestUtils.createInterfaceAddress("192.0.2.1", "255.255.0.0", (short) 24);
+        final InterfaceAddress address = TestUtils.createInterfaceAddress("192.0.2.1", "255.255.0.0", 24);
         doReturn(address).when(receiver).getInterfaceAddress();
         final NotifyListener listener = mock(NotifyListener.class);
         receiver.setNotifyListener(listener);
@@ -113,7 +113,7 @@ public class SsdpNotifyReceiverTest {
     @Test
     public void onReceive_ByeByeパケットは通知する() throws Exception {
         final SsdpNotifyReceiver receiver = spy(new SsdpNotifyReceiver(Address.IP_V4, NetworkUtils.getAvailableInet4Interfaces().get(0)));
-        final InterfaceAddress address = TestUtils.createInterfaceAddress("192.0.2.1", "255.255.0.0", (short) 24);
+        final InterfaceAddress address = TestUtils.createInterfaceAddress("192.0.2.1", "255.255.0.0", 24);
         doReturn(address).when(receiver).getInterfaceAddress();
         final NotifyListener listener = mock(NotifyListener.class);
         receiver.setNotifyListener(listener);
@@ -128,7 +128,7 @@ public class SsdpNotifyReceiverTest {
     @Test
     public void onReceive_LocationとSourceが不一致のメッセージは無視する() throws Exception {
         final SsdpNotifyReceiver receiver = spy(new SsdpNotifyReceiver(Address.IP_V4, NetworkUtils.getAvailableInet4Interfaces().get(0)));
-        final InterfaceAddress address = TestUtils.createInterfaceAddress("192.0.2.1", "255.255.0.0", (short) 24);
+        final InterfaceAddress address = TestUtils.createInterfaceAddress("192.0.2.1", "255.255.0.0", 24);
         doReturn(address).when(receiver).getInterfaceAddress();
         final NotifyListener listener = mock(NotifyListener.class);
         receiver.setNotifyListener(listener);
@@ -142,7 +142,7 @@ public class SsdpNotifyReceiverTest {
     @Test
     public void onReceive_IOExceptionが発生してもクラッシュしない() throws Exception {
         final SsdpNotifyReceiver receiver = spy(new SsdpNotifyReceiver(Address.IP_V4, NetworkUtils.getAvailableInet4Interfaces().get(0)));
-        final InterfaceAddress address = TestUtils.createInterfaceAddress("192.0.2.1", "255.255.0.0", (short) 24);
+        final InterfaceAddress address = TestUtils.createInterfaceAddress("192.0.2.1", "255.255.0.0", 24);
         doReturn(address).when(receiver).getInterfaceAddress();
         doThrow(new IOException()).when(receiver).createSsdpRequestMessage(ArgumentMatchers.any(byte[].class), anyInt());
         final NotifyListener listener = mock(NotifyListener.class);
@@ -169,22 +169,22 @@ public class SsdpNotifyReceiverTest {
     public void isSameSegment() throws Exception {
         assertThat(
                 SsdpNotifyReceiver.isSameSegment(
-                        TestUtils.createInterfaceAddress("192.168.0.1", "255.255.255.0", (short) 24),
+                        TestUtils.createInterfaceAddress("192.168.0.1", "255.255.255.0", 24),
                         InetAddress.getByName("192.168.0.255")),
                 is(true));
         assertThat(
                 SsdpNotifyReceiver.isSameSegment(
-                        TestUtils.createInterfaceAddress("192.168.0.1", "255.255.255.128", (short) 25),
+                        TestUtils.createInterfaceAddress("192.168.0.1", "255.255.255.128", 25),
                         InetAddress.getByName("192.168.0.255")),
                 is(false));
         assertThat(
                 SsdpNotifyReceiver.isSameSegment(
-                        TestUtils.createInterfaceAddress("192.168.0.1", "255.255.255.0", (short) 24),
+                        TestUtils.createInterfaceAddress("192.168.0.1", "255.255.255.0", 24),
                         InetAddress.getByName("192.168.1.255")),
                 is(false));
         assertThat(
                 SsdpNotifyReceiver.isSameSegment(
-                        TestUtils.createInterfaceAddress("192.168.0.1", "255.255.254.0", (short) 23),
+                        TestUtils.createInterfaceAddress("192.168.0.1", "255.255.254.0", 23),
                         InetAddress.getByName("192.168.1.255")),
                 is(true));
     }
